# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Workshop Finance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      parameters:
        - { name: workerId, in: query, schema: { type: integer } }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Expense" }
    post:
      operationId: createExpense
      tags: [expenses]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateExpenseBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Expense" }
  /expenses/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    delete:
      operationId: deleteExpense
      tags: [expenses]
      responses:
        "204": { description: Deleted }
...
  /analytics/summary:
    get:
      operationId: getSummary
      tags: [analytics]
      summary: Overall financial summary
      parameters:
        - { name: period, in: query, schema: { type: string, enum: [week, month, all] } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Summary" }
  /analytics/by-worker:
    get:
      operationId: getByWorker
      tags: [analytics]
      summary: Per-worker earnings & expenses
      parameters:
        - { name: period, in: query, schema: { type: string, enum: [week, month, all] } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WorkerStats" }
  /analytics/timeseries:
    get:
      operationId: getTimeseries
      tags: [analytics]
      summary: Income/expense over time
      parameters:
        - { name: bucket, in: query, schema: { type: string, enum: [day, week, month] } }
        - { name: period, in: query, schema: { type: string, enum: [week, month, all] } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TimeBucket" }
...
components:
  schemas:
...
    Expense:
      type: object
      required: [id, description, amount, category, occurredAt]
      properties:
        id: { type: integer }
        description: { type: string }
        amount: { type: number }
        category: { type: string, description: "e.g. rent, utilities, supplies, advance, other" }
        workerId: { type: integer, nullable: true, description: "Optional - tied to a specific worker (e.g. advance/loan)" }
        workerName: { type: string, nullable: true }
        paidWith: { type: string, enum: [cash, card] }
        occurredAt: { type: string, format: date-time }
    CreateExpenseBody:
      type: object
      required: [description, amount, category, paidWith]
      properties:
        description: { type: string }
        amount: { type: number }
        category: { type: string }
        workerId: { type: integer, nullable: true }
        paidWith: { type: string, enum: [cash, card] }
        occurredAt: { type: string, format: date-time }
...
    Summary:
      type: object
      required:
        - grossIncome
        - cardFees
        - netIncome
        - workerPayouts
        - workshopRevenue
        - generalExpenses
        - partsCost
        - workshopProfit
        - jobCount
        - cashIncome
        - cardIncome
      properties:
        grossIncome: { type: number }
        cardFees: { type: number }
        netIncome: { type: number }
        workerPayouts: { type: number }
        workshopRevenue: { type: number }
        generalExpenses: { type: number }
        partsCost: { type: number }
        workshopProfit: { type: number }
        jobCount: { type: integer }
        cashIncome: { type: number }
        cardIncome: { type: number }
    WorkerStats:
      type: object
      required: [workerId, name, jobCount, grossIncome, workerEarned, workshopEarned, expenses, netOwed]
      properties:
        workerId: { type: integer }
        name: { type: string }
        jobCount: { type: integer }
        grossIncome: { type: number }
        workerEarned: { type: number }
        workshopEarned: { type: number }
        expenses: { type: number, description: "Expenses tied to this worker (advances/loans)" }
        netOwed: { type: number, description: "What worker should receive at end of period" }
    TimeBucket:
      type: object
      required: [bucket, income, expenses, profit]
      properties:
        bucket: { type: string }
        income: { type: number }
        expenses: { type: number }
        profit: { type: number }
...