# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Workshop Finance API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpensePaidWith } from "./expensePaidWith";

export interface Expense {
  id: number;
  description: string;
  amount: number;
  /** e.g. rent, utilities, supplies, advance, other */
  category: string;
  /** Optional - tied to a specific worker (e.g. advance/loan) */
  workerId?: number | null;
  workerName?: string | null;
  paidWith?: ExpensePaidWith;
  occurredAt: Date;
}