# lib/api-zod/src/generated/types/workerStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Workshop Finance API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkerStats {
  workerId: number;
  name: string;
  jobCount: number;
  grossIncome: number;
  workerEarned: number;
  workshopEarned: number;
  /** Expenses tied to this worker (advances/loans) */
  expenses: number;
  /** What worker should receive at end of period */
  netOwed: number;
}